requirements.txt
Django
Pillow
psycopg2
django-phonenumber-field[phonenumbers]